subversion: Do not install perllocal.pod and *.la files
Perl build system(s) keep a list of installed modules in perllocal.pod
file.  In Clear we don't use that, and we clean it up as part of the
install step.

Since this package installs its perl modules after install step, it
misses the regular clean, so we clean it manually.